Add a batch mode in both tabs
Add a checkbox to activate batch mode.
Add a field for a folder containing models to convert.
Both tab.
No default values.
Add some basic logging into the console, kind of blind right now, wondering if it's dead or alive when converting to onnx, and if it's done or not.